fixed rsync

+++ machines/pb/reset_marmot.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 killall marmot
-rm /pb/marmot.cbor
+if [ -f /pb/marmot.cbor ]; then
+    rm /pb/marmot.cbor
+fi
 # /marmot -config /pb/marmot.toml -cleanup
 /marmot -config /pb/marmot.toml >> /pb/marmot.txt 2>&1 &
+touch /pb/marmot.active
+echo "Marmot started `date`" >> /pb/log.txt
+echo "Marmot started `date`"
+
